pyproject.toml
[project]
name = "n2n-redactor"
version = "0.0.1"
description = "Local-first PAN redaction CLI + API."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
dependencies = [
    "pdfplumber>=0.11.0",
    "PyMuPDF>=1.23.0",
    "typer>=0.9.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.30.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "ultralytics>=8.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]
paddle = ["paddleocr>=2.8.0"]
easy = ["easyocr>=1.7.1"]

[project.scripts]
n2n = "n2n.cli:app"

[tool.setuptools]
packages = ["n2n"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["_legacy", "build", "dist"]

[build-system]
requires = ["setuptools>=67.0", "wheel"]
build-backend = "setuptools.build_meta"
